Minio Delete upd

+++ src/main/java/ru/pirum1ch/cloudsave/services/MinioFileService.java
@@ -1,5 +1,6 @@
 package ru.pirum1ch.cloudsave.services;
 
+import io.jsonwebtoken.ExpiredJwtException;
 import io.minio.errors.*;
 import jakarta.persistence.PersistenceException;
 import lombok.extern.log4j.Log4j2;
@@ -68,21 +69,21 @@ public class MinioFileService {
 
     //TODO How to upload a few files
     @Transactional(rollbackFor = {IOException.class, MinioException.class, NoSuchAlgorithmException.class, InvalidKeyException.class})
-    public File upload(String fileName, MultipartFile file) throws IOException, MinioException, NoSuchAlgorithmException, InvalidKeyException {
+    public File upload(String fileName, MultipartFile file)
+            throws IOException, MinioException, NoSuchAlgorithmException, InvalidKeyException {
         log.log(Level.INFO, "Проверяем что входящие аргументы не пустые");
         if (fileName.isEmpty() || file.isEmpty()) {
             throw new IllegalArgumentException("Переданные на вход значения пусты!");
         }
+
         String key = fileManager.generateKey(file.getName());
-        log.log(Level.INFO, "Создали новый ключ для файла: " + key);
+
         if (fileRepo.findByName(fileName) != null) {
             log.log(Level.INFO, "Файл с таким именем уже есть");
-//            fileName = fileName + "_" + new SimpleDateFormat("HH:mm:ss").format(new Date().getTime());
-            fileName = new StringBuilder(fileName).
-                    insert(
-                            fileName.indexOf("."),
-                            "_" + new SimpleDateFormat("HH:mm:ss")
-                                    .format(new Date().getTime())).toString();
+            fileName = new StringBuilder(fileName)
+                    .insert(fileName.indexOf("."), "_" + new SimpleDateFormat("HH:mm:ss")
+                            .format(new Date().getTime()))
+                    .toString();
         }
 
         File uploadedFile = File.builder()
@@ -92,13 +93,12 @@ public class MinioFileService {
                 .extention(file.getContentType())
                 .uploadDate(new Date())
                 .build();
-
         log.log(Level.INFO, "Создали новый объект файла: \n" + uploadedFile);
+
         fileRepo.save(uploadedFile);
-
         log.log(Level.INFO, "Сохранили данные сущности в БД");
-        fileManager.minioUpload(file, key);
 
+        fileManager.minioUpload(file, key);
         return uploadedFile;
     }
 
@@ -136,51 +136,32 @@ public class MinioFileService {
         return foundFile;
     }
 
-    @Transactional(rollbackFor = {IOException.class,
-            FileNotFoundException.class,
-            ServerException.class,
-            InsufficientDataException.class,
-            ErrorResponseException.class,
-            NoSuchAlgorithmException.class,
-            InvalidKeyException.class,
-            InvalidResponseException.class,
-            XmlParserException.class,
-            InternalException.class,})
-    public void deleteFile(String filename) throws IOException, ServerException, InsufficientDataException, ErrorResponseException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
-        File file = null;
-        log.log(Level.INFO, "Проверяем что аргументы валидны.");
+    @Transactional(rollbackFor = {Exception.class})
+    public void deleteFile(String filename) throws ExpiredJwtException, IOException,
+            ServerException,
+            InsufficientDataException,
+            ErrorResponseException,
+            NoSuchAlgorithmException,
+            InvalidKeyException,
+            InvalidResponseException,
+            XmlParserException,
+            InternalException
+    {
+        log.log(Level.INFO, "Проверяем что входящие аргументы не пусты: " + filename);
         if (filename.isEmpty()) {
             log.log(Level.INFO, "Входящий аргумент пустой! Введите имя файла для удаления.");
             throw new IllegalArgumentException();
         }
 
-//        try {
-        //Ищем файл и передаем его в объект File
-//            try {
-                file = fileRepo.findByName(filename);
-                if (file == null || file.getSize() == 0) {
-                    log.log(Level.INFO, "Такого файла не существует! ");
-                    throw new FileNotFoundException();
-                }
-                log.log(Level.INFO, "Найден файл для удаления:\n" + file);
-//            } catch (IOException | NullPointerException exception) {
-//                log.log(Level.ERROR, exception.getLocalizedMessage());
-//            }
+        File file = fileRepo.findByName(filename);
+        if (fileRepo.findByName(filename) == null ) {
+            log.log(Level.INFO, "Файла "+ filename +" не существует! ");
+            throw new FileNotFoundException();
+        }
+        log.log(Level.INFO, "Найден файл для удаления:\n" + file.toString());
 
-            fileManager.deleteFile(file);
-            fileRepo.delete(file);
-            log.log(Level.INFO, "Файл удален.");
-//        } catch (IOException |
-//                 ServerException |
-//                 InsufficientDataException |
-//                 ErrorResponseException |
-//                 NoSuchAlgorithmException |
-//                 InvalidKeyException |
-//                 InvalidResponseException |
-//                 XmlParserException |
-//                 InternalException exception) {
-//            log.log(Level.ERROR, "Ошибка удаления файла:\n" + exception.getLocalizedMessage());
-//        }
-
+        fileManager.deleteFile(file);
+        fileRepo.delete(file);
+        log.log(Level.INFO, "Файл удален.");
     }
 }
